Changed Search Surah to DashBoard
