+++ app/build.gradle
@@ -10,7 +10,7 @@ android {
         minSdk 26
         targetSdk 33
         versionCode 2
-        versionName "2.3.1"
+        versionName "2.5"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
