+++ app/release/output-metadata.json
@@ -12,7 +12,7 @@
       "filters": [],
       "attributes": [],
       "versionCode": 2,
-      "versionName": "2.3.1",
+      "versionName": "2.5",
       "outputFile": "app-release.apk"
     }
   ],
