+++ app/src/main/res/layout/fragment_qirat.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="5dp"
-    tools:context=".Fragment.Qiraat">
+    tools:context=".Fragment.Qirat_Frag">
 
     <!-- <androidx.recyclerview.widget.RecyclerView
          android:id="@+id/RecyclerView"
