--- app/src/main/res/layout/searchsample.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_margin="5dp"
-    android:clickable="true"
-    android:foreground="?android:selectableItemBackgroundBorderless">
-
-    <TextView
-        android:id="@+id/surahIcon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:text="۝"
-        android:textSize="35sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/SearchSurahNumber"
-        android:layout_width="32dp"
-        android:layout_height="wrap_content"
-        android:background="@color/transparent"
-        android:text="01"
-        android:textAlignment="center"
-        android:textSize="17sp"
-        app:layout_constraintBottom_toBottomOf="@+id/surahIcon"
-        app:layout_constraintEnd_toEndOf="@+id/surahIcon"
-        app:layout_constraintHorizontal_bias="0.49"
-        app:layout_constraintStart_toStartOf="@+id/surahIcon"
-        app:layout_constraintTop_toTopOf="@+id/surahIcon"
-        app:layout_constraintVertical_bias="0.39"
-        tools:ignore="MissingConstraints" />
-
-    <TextView
-        android:id="@+id/SearchSurahName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="32dp"
-        android:text="Al-Fatiah"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="@+id/surahIcon"
-        app:layout_constraintStart_toEndOf="@+id/surahIcon"
-        app:layout_constraintTop_toTopOf="@+id/surahIcon" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
